fixed admin não poder excluir se usuário tiver eventos

+++ server/src/services/AdminService.ts
@@ -51,12 +51,29 @@ export async function deleteUserAdminService (userId: string, username: string) 
     if(admin?.username !== "admin") {
         throw new Error('Usuário não é admin');
     } else {
+
+        const deleteUser = await prisma.user.findUnique({
+            where: {
+                username: username
+            }
+        })
+
+        if(!deleteUser) {
+            throw new Error('Usuário não para ser deletado não existe');
+        }
+
+        const eventsDeleted = await prisma.event.deleteMany({
+            where: {
+                userId: deleteUser.id
+            }
+        })
+
        const user = await prisma.user.delete({
         where: {
             username: username
         }
        })
     
-        return user
+        return {user, eventsDeleted}
     }
 }
